% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal.R
\name{crossVal}
\alias{crossVal}
\title{Five fold cross-validation on the EEG data}
\usage{
crossVal(
  seed = 113,
  Nmcmc = NULL,
  burnin = NULL,
  thin = NULL,
  v0 = 0.005,
  a1 = 5,
  a2 = 50,
  q = 10,
  nCores = 1
)
}
\arguments{
\item{seed}{seed for creating splits, defaults to 113}

\item{Nmcmc}{number of MCMC samples to generate}

\item{burnin}{burnin samples}

\item{thin}{thinning samples \eqn{>0}}

\item{v0}{hyperparameter v0, defaults to 0.005}

\item{a1}{hyperparameter a1, defaults to 5}

\item{a2}{hyperparameter a2, defaults to 50}

\item{q}{hyperparameter q, defaults to 10}

\item{nCores}{number of cores to use to run cross validation parallelly, defaults to 1}
}
\description{
Five fold cross-validation by specifying a specific split by seed
}
\examples{
crossVal()
}
\keyword{crossVal}
