% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.mat_eeg.R
\docType{data}
\name{dist.mat_eeg}
\alias{dist.mat_eeg}
\title{Distance Matrix}
\format{
A matrix of dimension 57x57.
}
\usage{
dist.mat_eeg
}
\description{
This is the distance matrix for the 57 locations whose coordinates
are known.
}
\keyword{datasets}
