% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind64to57_eeg.R
\docType{data}
\name{ind64to57_eeg}
\alias{ind64to57_eeg}
\title{Indexing 64 locations to 57}
\format{
A vector of integers of length 57.
}
\usage{
ind64to57_eeg
}
\description{
This vector indexes the the 57 locations which have known
coordinates from the 64 locations in the EEG data.
}
\keyword{datasets}
