% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strucBayes.R
\name{strucBayes}
\alias{strucBayes}
\title{Local Bayesian modeling}
\usage{
strucBayes(
  y = NULL,
  X = NULL,
  dist.mat = NULL,
  Nmcmc = NULL,
  burnin = NULL,
  thin = NULL,
  X.test = NULL,
  v0 = 0.005,
  a1 = 5,
  a2 = 50,
  q = 10
)
}
\arguments{
\item{y}{binary response \eqn{n} values}

\item{X}{data tensor \eqn{n x L x \tau}}

\item{dist.mat}{distance matrix between \eqn{L} locations}

\item{Nmcmc}{number of MCMC samples to generate}

\item{burnin}{burnin samples}

\item{thin}{thinning samples \eqn{>0}}

\item{X.test}{test data \eqn{m x p x tau}, defaults to NULL}

\item{v0}{hyperparameter \eqn{v_0}, defaults to 0.005}

\item{a1}{hyperparameter \eqn{a_1}, defaults to 5}

\item{a2}{hyperparameter \eqn{a_2}, defaults to 50}

\item{q}{hyperparameter \eqn{q}, defaults to 10}
}
\description{
Local Bayesian modeling for the EEG data with structured spike and slab prior. Performs variable selection and prediction for local modeling using structured spike and slab prior.
}
\examples{
strucBayes()
}
\keyword{strucBayes}
