% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strucPriorT.R
\name{strucPriorT}
\alias{strucPriorT}
\title{MCMC algorithm for the estimation of local model}
\usage{
strucPriorT(
  y,
  X,
  c_init,
  beta_init,
  zeta_init,
  nusq_init,
  l_init,
  prior.mu,
  prior.Delta,
  v0,
  a1,
  a2,
  q,
  Nmcmc,
  ind
)
}
\arguments{
\item{y}{binary response n values}

\item{X}{data at time \eqn{t} - dimension \eqn{n x L}}

\item{c_init}{initial value for \eqn{c}}

\item{beta_init}{initial values for \eqn{\beta}}

\item{zeta_init}{initial values for \eqn{\zeta}}

\item{nusq_init}{initial values for \eqn{\nu^2}}

\item{l_init}{initial value for \eqn{\lambda}}

\item{prior.mu}{prior mean for \eqn{\lambda}}

\item{prior.Delta}{prior covariance for \eqn{\lambda}}

\item{v0}{hyperparameter \eqn{v0}}

\item{a1}{hyperparameter \eqn{a1}}

\item{a2}{hyperparameter \eqn{a2}}

\item{q}{hyperparameter \eqn{q}}

\item{Nmcmc}{number of MCMC samples to generate}

\item{ind}{indices of MCMC samples to use}
}
\description{
Gibbs sampling, and slice sampling within Gibbs sampling for a local model
}
\examples{
strucPriorT()
}
\keyword{strucPriorT()}
