% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_eeg.R
\docType{data}
\name{X_eeg}
\alias{X_eeg}
\title{EEG Data}
\format{
A large array of dimensions 256x64x122.
}
\usage{
X_eeg
}
\description{
This array is the EEG data for 122 subjects at 64 locations.
Measurements are taken at 256 time points.
}
\keyword{datasets}
