% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_eeg.R
\docType{data}
\name{y_eeg}
\alias{y_eeg}
\title{Response Vector}
\format{
A vector with n=122 entries indicating alcoholics and controls.
}
\usage{
y_eeg
}
\description{
The response vector indicating the alcoholic status of the patient
}
\keyword{datasets}
